Move DemoControls outside Canvas
