+++ AGENTS.md
@@ -80,6 +80,7 @@ This file tracks Codex progress and upcoming tasks. Keep it chronological and ap
 - **57** – Centered ripple zoom on foreground position with new uniform and limited range. Lint and build pass.
 - **58** – Added "center zoom" toggle to switch ripple zoom between canvas center and foreground. Lint and build pass.
 - **59** – Faded out new Overview panel on first drag and hid Tweakpane on small screens. Lint and build pass.
+- **60** – Moved DemoControls outside the R3F Canvas by hoisting state to `App` and updating component props. Lint and build pass.
 
 ## Next Steps
